@@ -1,26 +1,28 @@
 <script lang="ts">
-  import { Progressbar, Button } from 'flowbite-svelte';
-  import { ChevronUpSolid, ChevronDownSolid, CloseSolid } from 'flowbite-svelte-icons';
+  import { Button, Progressbar } from 'flowbite-svelte';
+  import { CaretDownSolid, CaretUpSolid, CloseCircleSolid } from 'flowbite-svelte-icons';
   import { formatBytes } from '$lib/util';
   import { progressTrackerStore } from '$lib/util/progress-tracker';
 
-  let expanded = true;
-  
+  let expanded = $state(true);
+
   function toggleExpanded() {
     expanded = !expanded;
   }
-  
+
   function removeProcess(id: string) {
     progressTrackerStore.removeProcess(id);
   }
 </script>
 
 {#if $progressTrackerStore.processes.length > 0}
-  <div class="fixed bottom-4 right-4 z-50 bg-white dark:bg-gray-800 rounded-lg shadow-lg border border-gray-200 dark:border-gray-700 w-80 overflow-hidden transition-all duration-200">
-    <div 
-      class="flex justify-between items-center p-3 bg-primary-100 dark:bg-primary-900 cursor-pointer" 
-      on:click={toggleExpanded}
-      on:keydown={(e) => e.key === 'Enter' && toggleExpanded()}
+  <div
+    class="fixed bottom-4 right-4 z-50 bg-white dark:bg-gray-800 rounded-lg shadow-lg border border-gray-200 dark:border-gray-700 w-80 overflow-hidden transition-all duration-200"
+  >
+    <div
+      class="flex justify-between items-center p-3 bg-primary-100 dark:bg-primary-900 cursor-pointer"
+      onclick={toggleExpanded}
+      onkeydown={(e) => e.key === 'Enter' && toggleExpanded()}
       role="button"
       tabindex="0"
     >
@@ -33,13 +35,13 @@
       </div>
       <div>
         {#if expanded}
-          <ChevronDownSolid class="w-4 h-4" />
+          <CaretDownSolid class="w-4 h-4" />
         {:else}
-          <ChevronUpSolid class="w-4 h-4" />
+          <CaretUpSolid class="w-4 h-4" />
         {/if}
       </div>
     </div>
-    
+
     {#if expanded}
       <div class="p-3 max-h-96 overflow-y-auto">
         {#each $progressTrackerStore.processes as process (process.id)}
@@ -47,21 +49,21 @@
             <div class="flex justify-between items-center mb-1">
               <div class="text-sm font-medium">{process.description}</div>
               <Button size="xs" color="none" class="p-1" on:click={() => removeProcess(process.id)}>
-                <CloseSolid class="w-3 h-3" />
+                <CloseCircleSolid class="w-3 h-3" />
               </Button>
             </div>
-            
+
             {#if process.status}
               <div class="text-xs text-gray-600 dark:text-gray-400 mb-1">{process.status}</div>
             {/if}
-            
+
             <div class="flex justify-between text-xs mb-1">
               <span>{Math.round(process.progress)}%</span>
               {#if process.bytesLoaded !== undefined && process.totalBytes !== undefined}
                 <span>{formatBytes(process.bytesLoaded)} / {formatBytes(process.totalBytes)}</span>
               {/if}
             </div>
-            
+
             <Progressbar progress={process.progress.toString()} size="h-2" />
           </div>
         {/each}
