@@ -1,11 +1,24 @@
 <script lang="ts">
+  import { run } from 'svelte/legacy';
+
   import { currentSeries, currentVolume, currentVolumeData } from '$lib/catalog';
-  import { Panzoom, panzoomStore, toggleFullScreen, zoomDefault, zoomFitToScreen } from '$lib/panzoom';
+  import {
+    Panzoom,
+    panzoomStore,
+    toggleFullScreen,
+    zoomDefault,
+    zoomFitToScreen
+  } from '$lib/panzoom';
   import { progress, settings, updateProgress, type VolumeSettings } from '$lib/settings';
   import { clamp, debounce, fireExstaticEvent } from '$lib/util';
   import { Input, Popover, Range, Spinner } from 'flowbite-svelte';
   import MangaPage from './MangaPage.svelte';
-  import { ChervonDoubleLeftSolid, ChervonDoubleRightSolid, ChevronLeftSolid, ChevronRightSolid } from 'flowbite-svelte-icons';
+  import {
+    BackwardStepSolid,
+    CaretLeftSolid,
+    CaretRightSolid,
+    ForwardStepSolid
+  } from 'flowbite-svelte-icons';
   import Cropper from './Cropper.svelte';
   import SettingsButton from './SettingsButton.svelte';
   import { getCharCount } from '$lib/util/count-chars';
@@ -14,18 +27,11 @@
   import { onMount } from 'svelte';
 
   // TODO: Refactor this whole mess
-  export let volumeSettings: VolumeSettings;
-  $: volume = $currentVolume;
-  $: volumeData = $currentVolumeData;
-  $: pages = volumeData?.pages || [];
+  interface Props {
+    volumeSettings: VolumeSettings;
+  }
 
-  $: page = $progress?.[volume?.volume_uuid || 0] || 1;
-  $: index = page - 1;
-  $: navAmount =
-    volumeSettings.singlePageView ||
-    (volumeSettings.hasCover && !volumeSettings.singlePageView && index === 0)
-      ? 1
-      : 2;
+  let { volumeSettings }: Props = $props();
 
   let start: Date;
 
@@ -56,11 +62,14 @@
       const { charCount } = getCharCount(pages, pageClamped);
       if (pageClamped !== newPage) {
         let seriesVolumes = $currentSeries;
-        const currentVolumeIndex = seriesVolumes.findIndex((v) => v.volume_uuid === volume.volume_uuid);
+        const currentVolumeIndex = seriesVolumes.findIndex(
+          (v) => v.volume_uuid === volume.volume_uuid
+        );
         if (newPage < 1) {
           // open previous volume
           const previousVolume = seriesVolumes[currentVolumeIndex - 1];
-          if (previousVolume) window.location.href = `/${volume.series_uuid}/${previousVolume.volume_uuid}`;
+          if (previousVolume)
+            window.location.href = `/${volume.series_uuid}/${previousVolume.volume_uuid}`;
           else window.location.href = `/${volume.series_uuid}`;
         } else if (newPage > pages.length) {
           // open next volume
@@ -79,29 +88,6 @@
       }
     }
   }
-
-  $: showSecondPage = () => {
-    if (!pages) {
-      return false;
-    }
-
-    if (volumeSettings.singlePageView || index + 1 >= pages.length) {
-      return false;
-    }
-
-    if (index === 0 && volumeSettings.hasCover) {
-      return false;
-    }
-
-    return true;
-  };
-
-  $: manualPage = page;
-  $: pageDisplay = showSecondPage()
-    ? `${page},${page + 1} / ${pages?.length}`
-    : `${page} / ${pages?.length}`;
-
-  $: charDisplay = `${charCount} / ${maxCharCount}`;
 
   function onInputClick(this: any) {
     this.select();
@@ -153,10 +139,6 @@
         break;
     }
   }
-
-  $: charCount = $settings.charCount ? getCharCount(pages, page).charCount : 0;
-  $: maxCharCount = getCharCount(pages).charCount;
-  $: totalLineCount = getCharCount(pages).lineCount;
 
   let startX = 0;
   let startY = 0;
@@ -213,23 +195,6 @@
     }
   }
 
-  $: {
-    if (volume) {
-      const { charCount, lineCount } = getCharCount(pages, page);
-
-      fireExstaticEvent('mokuro-reader:page.change', {
-        title: volume.series_title,
-        volumeName: volume.volume_title,
-        currentCharCount: charCount,
-        currentPage: page,
-        totalPages: pages.length,
-        totalCharCount: maxCharCount || 0,
-        currentLineCount: lineCount,
-        totalLineCount
-      });
-    }
-  }
-
   onMount(() => {
     if ($settings.defaultFullscreen) {
       document.documentElement.requestFullscreen();
@@ -256,13 +221,66 @@
       });
     }
   });
+  let volume = $derived($currentVolume);
+  let volumeData = $derived($currentVolumeData);
+  let pages = $derived(volumeData?.pages || []);
+  let page = $derived($progress?.[volume?.volume_uuid || 0] || 1);
+  let index = $derived(page - 1);
+  let navAmount = $derived(
+    volumeSettings.singlePageView ||
+      (volumeSettings.hasCover && !volumeSettings.singlePageView && index === 0)
+      ? 1
+      : 2
+  );
+  let showSecondPage = $derived(() => {
+    if (!pages) {
+      return false;
+    }
+
+    if (volumeSettings.singlePageView || index + 1 >= pages.length) {
+      return false;
+    }
+
+    if (index === 0 && volumeSettings.hasCover) {
+      return false;
+    }
+
+    return true;
+  });
+  let manualPage = $state(0);
+  run(() => {
+    manualPage = page;
+  });
+  let pageDisplay = $derived(
+    showSecondPage() ? `${page},${page + 1} / ${pages?.length}` : `${page} / ${pages?.length}`
+  );
+  let charCount = $derived($settings.charCount ? getCharCount(pages, page).charCount : 0);
+  let maxCharCount = $derived(getCharCount(pages).charCount);
+  let charDisplay = $derived(`${charCount} / ${maxCharCount}`);
+  let totalLineCount = $derived(getCharCount(pages).lineCount);
+  run(() => {
+    if (volume) {
+      const { charCount, lineCount } = getCharCount(pages, page);
+
+      fireExstaticEvent('mokuro-reader:page.change', {
+        title: volume.series_title,
+        volumeName: volume.volume_title,
+        currentCharCount: charCount,
+        currentPage: page,
+        totalPages: pages.length,
+        totalCharCount: maxCharCount || 0,
+        currentLineCount: lineCount,
+        totalLineCount
+      });
+    }
+  });
 </script>
 
 <svelte:window
-  on:resize={zoomDefault}
-  on:keyup={handleShortcuts}
-  on:touchstart={handleTouchStart}
-  on:touchend={handlePointerUp}
+  onresize={zoomDefault}
+  onkeyup={handleShortcuts}
+  ontouchstart={handleTouchStart}
+  ontouchend={handlePointerUp}
 />
 <svelte:head>
   <title>{volume?.volume_title || 'Volume'}</title>
@@ -279,16 +297,12 @@
   <Popover placement="bottom" trigger="click" triggeredBy="#page-num" class="z-20 w-full max-w-xs">
     <div class="flex flex-col gap-3">
       <div class="flex flex-row items-center gap-5 z-10">
-        <ChervonDoubleLeftSolid
-          on:click={() => changePage(volumeSettings.rightToLeft ? pages.length : 1, true)}
-          class="hover:text-primary-600"
-          size="sm"
-        />
-        <ChevronLeftSolid
-          on:click={(e) => left(e, true)}
-          class="hover:text-primary-600"
-          size="sm"
-        />
+        <button onclick={() => changePage(volumeSettings.rightToLeft ? pages.length : 1, true)}>
+          <BackwardStepSolid class="hover:text-primary-600" size="sm" />
+        </button>
+        <button onclick={(e) => left(e, true)}>
+          <CaretLeftSolid class="hover:text-primary-600" size="sm" />
+        </button>
         <Input
           type="number"
           size="sm"
@@ -303,16 +317,12 @@
           }}
           on:blur={onManualPageChange}
         />
-        <ChevronRightSolid
-          on:click={(e) => right(e, true)}
-          class="hover:text-primary-600"
-          size="sm"
-        />
-        <ChervonDoubleRightSolid
-          on:click={() => changePage(volumeSettings.rightToLeft ? 1 : pages.length, true)}
-          class="hover:text-primary-600"
-          size="sm"
-        />
+        <button onclick={(e) => right(e, true)}>
+          <CaretRightSolid class="hover:text-primary-600" size="sm" />
+        </button>
+        <button onclick={() => changePage(volumeSettings.rightToLeft ? 1 : pages.length, true)}>
+          <ForwardStepSolid class="hover:text-primary-600" size="sm" />
+        </button>
       </div>
       <div style:direction={volumeSettings.rightToLeft ? 'rtl' : 'ltr'}>
         <Range
@@ -334,30 +344,30 @@
       <button
         class="h-full fixed -left-full z-10 w-full hover:bg-slate-400 opacity-[0.01]"
         style:margin-left={`${$settings.edgeButtonWidth}px`}
-        on:mousedown={mouseDown}
-        on:mouseup={left}
-      />
+        onmousedown={mouseDown}
+        onmouseup={left}
+      ></button>
       <button
         class="h-full fixed -right-full z-10 w-full hover:bg-slate-400 opacity-[0.01]"
         style:margin-right={`${$settings.edgeButtonWidth}px`}
-        on:mousedown={mouseDown}
-        on:mouseup={right}
-      />
+        onmousedown={mouseDown}
+        onmouseup={right}
+      ></button>
       <button
         class="h-screen fixed top-full -left-full z-10 w-[150%] hover:bg-slate-400 opacity-[0.01]"
-        on:mousedown={mouseDown}
-        on:mouseup={left}
-      />
+        onmousedown={mouseDown}
+        onmouseup={left}
+      ></button>
       <button
         class="h-screen fixed top-full -right-full z-10 w-[150%] hover:bg-slate-400 opacity-[0.01]"
-        on:mousedown={mouseDown}
-        on:mouseup={right}
-      />
+        onmousedown={mouseDown}
+        onmouseup={right}
+      ></button>
       <div
         class="flex flex-row"
         class:flex-row-reverse={!volumeSettings.rightToLeft}
         style:filter={`invert(${$settings.invertColors ? 1 : 0})`}
-        on:dblclick={onDoubleTap}
+        ondblclick={onDoubleTap}
         role="none"
         id="manga-panel"
       >
@@ -372,17 +382,17 @@
   </div>
   {#if !$settings.mobile}
     <button
-      on:mousedown={mouseDown}
-      on:mouseup={left}
+      onmousedown={mouseDown}
+      onmouseup={left}
       class="left-0 top-0 absolute h-full w-16 hover:bg-slate-400 opacity-[0.01]"
       style:width={`${$settings.edgeButtonWidth}px`}
-    />
+    ></button>
     <button
-      on:mousedown={mouseDown}
-      on:mouseup={right}
+      onmousedown={mouseDown}
+      onmouseup={right}
       class="right-0 top-0 absolute h-full w-16 hover:bg-slate-400 opacity-[0.01]"
       style:width={`${$settings.edgeButtonWidth}px`}
-    />
+    ></button>
   {/if}
 {:else}
   <div class="fixed z-50 left-1/2 top-1/2">
