@@ -4,48 +4,54 @@
   import { settings } from '$lib/settings';
   import { imageToWebp, showCropper, updateLastCard } from '$lib/anki-connect';
 
-  export let page: Page;
-  export let src: File;
+  interface Props {
+    page: Page;
+    src: File;
+  }
 
-  $: textBoxes = page.blocks
-    .map((block) => {
-      const { img_height, img_width } = page;
-      const { box, font_size, lines, vertical } = block;
+  let { page, src }: Props = $props();
 
-      let [_xmin, _ymin, _xmax, _ymax] = box;
+  let textBoxes = $derived(
+    page.blocks
+      .map((block) => {
+        const { img_height, img_width } = page;
+        const { box, font_size, lines, vertical } = block;
 
-      const xmin = clamp(_xmin, 0, img_width);
-      const ymin = clamp(_ymin, 0, img_height);
-      const xmax = clamp(_xmax, 0, img_width);
-      const ymax = clamp(_ymax, 0, img_height);
+        let [_xmin, _ymin, _xmax, _ymax] = box;
 
-      const width = xmax - xmin;
-      const height = ymax - ymin;
-      const area = width * height;
+        const xmin = clamp(_xmin, 0, img_width);
+        const ymin = clamp(_ymin, 0, img_height);
+        const xmax = clamp(_xmax, 0, img_width);
+        const ymax = clamp(_ymax, 0, img_height);
 
-      const textBox = {
-        left: `${xmin}px`,
-        top: `${ymin}px`,
-        width: `${width}px`,
-        height: `${height}px`,
-        fontSize: $settings.fontSize === 'auto' ? `${font_size}px` : `${$settings.fontSize}pt`,
-        writingMode: vertical ? 'vertical-rl' : 'horizontal-tb',
-        lines,
-        area
-      };
+        const width = xmax - xmin;
+        const height = ymax - ymin;
+        const area = width * height;
 
-      return textBox;
-    })
-    .sort(({ area: a }, { area: b }) => {
-      return b - a;
-    });
+        const textBox = {
+          left: `${xmin}px`,
+          top: `${ymin}px`,
+          width: `${width}px`,
+          height: `${height}px`,
+          fontSize: $settings.fontSize === 'auto' ? `${font_size}px` : `${$settings.fontSize}pt`,
+          writingMode: vertical ? 'vertical-rl' : 'horizontal-tb',
+          lines,
+          area
+        };
 
-  $: fontWeight = $settings.boldFont ? 'bold' : '400';
-  $: display = $settings.displayOCR ? 'block' : 'none';
-  $: border = $settings.textBoxBorders ? '1px solid red' : 'none';
-  $: contenteditable = $settings.textEditable;
+        return textBox;
+      })
+      .sort(({ area: a }, { area: b }) => {
+        return b - a;
+      })
+  );
 
-  $: triggerMethod = $settings.ankiConnectSettings.triggerMethod || 'both';
+  let fontWeight = $derived($settings.boldFont ? 'bold' : '400');
+  let display = $derived($settings.displayOCR ? 'block' : 'none');
+  let border = $derived($settings.textBoxBorders ? '1px solid red' : 'none');
+  let contenteditable = $derived($settings.textEditable);
+
+  let triggerMethod = $derived($settings.ankiConnectSettings.triggerMethod || 'both');
 
   async function onUpdateCard(lines: string[]) {
     if ($settings.ankiConnectSettings.enabled) {
@@ -89,8 +95,8 @@
     style:border
     style:writing-mode={writingMode}
     role="none"
-    on:contextmenu={(e) => onContextMenu(e, lines)}
-    on:dblclick={(e) => onDoubleTap(e, lines)}
+    oncontextmenu={(e) => onContextMenu(e, lines)}
+    ondblclick={(e) => onDoubleTap(e, lines)}
     {contenteditable}
   >
     {#each lines as line}
