@@ -1,10 +1,10 @@
 <script lang="ts">
-  import { Select, Input, Label } from 'flowbite-svelte';
-  import { settings, updateSetting } from '$lib/settings';
+  import { Input, Label, Select } from 'flowbite-svelte';
   import type { SettingsKey } from '$lib/settings';
+  import { settings, updateSetting } from '$lib/settings';
 
-  $: zoomModeValue = $settings.zoomDefault;
-  $: fontSizeValue = $settings.fontSize;
+  let zoomModeValue = $derived($settings.zoomDefault);
+  let fontSizeValue = $derived($settings.fontSize);
 
   let zoomModes = [
     { value: 'zoomFitToScreen', name: 'Fit to screen' },
