@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { Drawer, CloseButton, Button, Accordion } from 'flowbite-svelte';
+  import { Accordion, Button, CloseButton, Drawer } from 'flowbite-svelte';
   import { UserSettingsSolid } from 'flowbite-svelte-icons';
   import { sineIn } from 'svelte/easing';
   import { resetSettings } from '$lib/settings';
@@ -21,7 +21,12 @@
     easing: sineIn
   };
 
-  export let hidden = true;
+  interface Props {
+    hidden?: boolean;
+  }
+
+  // In Svelte 5, we need to make sure the hidden prop is properly bindable
+  let { hidden = $bindable(true) }: Props = $props();
 
   function onReset() {
     hidden = true;
