@@ -2,7 +2,7 @@
   import { settings, updateVolumeDefaults, type VolumeDefaultsKey } from '$lib/settings';
   import { AccordionItem, Helper, Toggle } from 'flowbite-svelte';
 
-  $: toggles = [
+  let toggles = $derived([
     { key: 'rightToLeft', text: 'Right to left', value: $settings.volumeDefaults?.rightToLeft },
     {
       key: 'singlePageView',
@@ -10,11 +10,13 @@
       value: $settings.volumeDefaults?.singlePageView
     },
     { key: 'hasCover', text: 'First page is cover', value: $settings.volumeDefaults?.hasCover }
-  ] as { key: VolumeDefaultsKey; text: string; value: any }[];
+  ] as { key: VolumeDefaultsKey; text: string; value: any }[]);
 </script>
 
 <AccordionItem open>
-  <span slot="header">Volume defaults</span>
+  {#snippet header()}
+    <span>Volume defaults</span>
+  {/snippet}
   <div class="flex flex-col gap-5">
     <Helper>The default settings that are applied when you start a new volume</Helper>
     {#each toggles as { key, text, value }}
