@@ -1,16 +1,36 @@
 <script lang="ts">
-  import { Button, Dropzone, Modal, Spinner, Accordion, AccordionItem } from 'flowbite-svelte';
-  import FileUpload from './FileUpload.svelte';
-  import { processFiles } from '$lib/upload';
+  import { Accordion, AccordionItem, Button, Dropzone, Modal, Spinner } from 'flowbite-svelte';
+  import { processFiles, scanFiles } from '$lib/upload';
   import { onMount } from 'svelte';
-  import { scanFiles } from '$lib/upload';
   import { formatBytes } from '$lib/util/upload';
   import { toClipboard } from '$lib/util';
 
-  export let open = false;
+  interface Props {
+    open?: boolean;
+  }
 
-  let promise: Promise<void>;
-  let files: FileList | undefined = undefined;
+  // In Svelte 5, we need to make sure the open prop is properly bindable
+  let { open = $bindable(false) }: Props = $props();
+
+  let promise: Promise<void> = $state();
+  let files: FileList | undefined = $state(undefined);
+  let isMobileDevice = $state(false);
+
+  // Check if the device is mobile
+  function checkMobileDevice() {
+    // Using userAgent to detect mobile devices
+    const userAgent = navigator.userAgent || navigator.vendor || (window as any).opera;
+    if (/android|iPad|iPhone|iPod|webOS|BlackBerry|IEMobile|Opera Mini/i.test(userAgent)) {
+      return true;
+    }
+
+    // Additional check for touch devices and small screens
+    return (
+      'ontouchstart' in window ||
+      navigator.maxTouchPoints > 0 ||
+      (window.innerWidth <= 800 && window.innerHeight <= 900)
+    );
+  }
 
   async function onUpload() {
     if (files) {
@@ -29,9 +49,12 @@
     draggedFiles = undefined;
   }
 
-  let storageSpace = '';
+  let storageSpace = $state('');
 
   onMount(() => {
+    // Check if device is mobile
+    isMobileDevice = checkMobileDevice();
+
     navigator?.storage?.estimate().then(({ usage, quota }) => {
       if (usage && quota) {
         storageSpace = `Storage: ${formatBytes(usage)} / ${formatBytes(quota)}`;
@@ -40,9 +63,9 @@
   });
 
   let filePromises: Promise<File>[];
-  let draggedFiles: File[] | undefined;
-  let loading = false;
-  $: disabled = loading || (!draggedFiles && !files);
+  let draggedFiles: File[] | undefined = $state();
+  let loading = $state(false);
+  let disabled = $derived(loading || (!draggedFiles && !files));
 
   const dropHandle = async (event: DragEvent) => {
     loading = true;
@@ -83,7 +106,7 @@
   let highlightStyle =
     'flex flex-col justify-center items-center w-full h-64 bg-gray-50 rounded-lg border-2 border-gray-300 border-dashed cursor-pointer dark:bg-bray-800 dark:bg-gray-700 bg-gray-100 dark:border-gray-600 dark:border-gray-500 dark:bg-gray-600';
 
-  let activeStyle = defaultStyle;
+  let activeStyle = $state(defaultStyle);
 </script>
 
 <Modal title="Upload" bind:open outsideclose on:close={reset}>
@@ -93,14 +116,16 @@
   {:then}
     <Accordion flush>
       <AccordionItem>
-        <span slot="header">What to upload?</span>
+        {#snippet header()}
+          <span>What to upload?</span>
+        {/snippet}
         <div class="flex flex-col gap-5">
           <div>
             <p>
               Firstly, ensure that you process your manga with the <b>0.2.0-beta.6</b> of mokuro, you
               can install it by running the following command:
             </p>
-            <div role="none" on:click={toClipboard}>
+            <div role="none" onclick={toClipboard}>
               <code class="text-primary-600 bg-slate-900"
                 >pip3 install git+https://github.com/kha-white/mokuro.git@web-reader</code
               >
@@ -111,9 +136,15 @@
             manga along with the <code>.mokuro</code> files.
           </p>
           <p>
-            On mobile, uploading via directory is not supported so you will need to zip your manga
-            first and then upload it via
-            <code class="text-primary-600 bg-slate-900">choose files</code>.
+            {#if isMobileDevice}
+              <b>Note:</b> On mobile devices, directory upload is not supported. Please zip your
+              manga first and then upload it via
+              <code class="text-primary-600 bg-slate-900">choose files</code>.
+            {:else}
+              On mobile devices, directory upload is not supported. If you're using a mobile device,
+              you will need to zip your manga first and then upload it via
+              <code class="text-primary-600 bg-slate-900">choose files</code>.
+            {/if}
           </p>
         </div>
       </AccordionItem>
@@ -162,10 +193,51 @@
         <Spinner />
       {:else}
         <p class="mb-2 text-sm text-gray-500 dark:text-gray-400">
-          Drag and drop / <FileUpload bind:files accept=".mokuro,.zip,.cbz" multiple
-            >choose files</FileUpload
-          > /
-          <FileUpload bind:files webkitdirectory>choose directory</FileUpload>
+          Drag and drop /
+          <button
+            type="button"
+            class="text-primary-600 dark:text-primary-500 hover:underline bg-transparent border-none p-0 m-0 cursor-pointer inline-flex"
+            onclick={() => {
+              const input = document.createElement('input');
+              input.type = 'file';
+              input.accept = '.mokuro,.zip,.cbz';
+              input.multiple = true;
+              input.onchange = (e) => {
+                if (e.target.files.length > 0) {
+                  files = e.target.files;
+                }
+              };
+              input.click();
+            }}
+          >
+            choose files
+          </button>
+
+          {#if !isMobileDevice}
+            / <button
+              type="button"
+              class="text-primary-600 dark:text-primary-500 hover:underline bg-transparent border-none p-0 m-0 cursor-pointer inline-flex"
+              onclick={() => {
+                const input = document.createElement('input');
+                input.type = 'file';
+                input.setAttribute('webkitdirectory', '');
+                input.onchange = (e) => {
+                  if (e.target.files.length > 0) {
+                    files = e.target.files;
+                  }
+                };
+                input.click();
+              }}
+            >
+              choose directory
+            </button>
+          {/if}
+
+          {#if isMobileDevice}
+            <span class="ml-1 text-xs text-gray-500 dark:text-gray-400 italic">
+              (directory upload not available on mobile)
+            </span>
+          {/if}
         </p>
       {/if}
     </Dropzone>
