@@ -1,27 +1,31 @@
 <script lang="ts">
-  import { processFiles } from '$lib/upload';
-  import { parseVolumesFromJson } from '$lib/settings';
+  import { run } from 'svelte/legacy';
 
-  /** @type {string} */
-  export let accessToken = '';
-  import { formatBytes, showSnackbar, uploadFile } from '$lib/util';
+  import { processFiles } from '$lib/upload';
+  import { parseVolumesFromJson, profiles, volumes } from '$lib/settings';
+
+  import { promptConfirmation, showSnackbar, uploadFile } from '$lib/util';
   import { Button } from 'flowbite-svelte';
   import { onMount } from 'svelte';
-  import { promptConfirmation } from '$lib/util';
   import { GoogleSolid } from 'flowbite-svelte-icons';
-  import { profiles, volumes } from '$lib/settings';
   import { progressTrackerStore } from '$lib/util/progress-tracker';
-  
+
+  interface Props {
+    accessToken?: string;
+  }
+
+  let { accessToken = $bindable('') }: Props = $props();
+
   // Helper function to handle errors consistently
   function handleDriveError(error: any, context: string) {
     // Check if it's a connectivity issue
     const errorMessage = error.toString().toLowerCase();
-    const isConnectivityError = 
-      errorMessage.includes('network') || 
-      errorMessage.includes('connection') || 
+    const isConnectivityError =
+      errorMessage.includes('network') ||
+      errorMessage.includes('connection') ||
       errorMessage.includes('offline') ||
       errorMessage.includes('internet');
-    
+
     if (!isConnectivityError) {
       // Log the user out for non-connectivity errors
       logout();
@@ -29,7 +33,7 @@
     } else {
       showSnackbar('Connection error: Please check your internet connection');
     }
-    
+
     console.error(`${context} error:`, error);
   }
 
@@ -48,15 +52,17 @@
 
   let tokenClient: any;
   let readerFolderId = '';
-  let volumeDataId = '';
-  let profilesId = '';
+  let volumeDataId = $state('');
+  let profilesId = $state('');
 
   // This variable is used to track if we're connected to Google Drive
   // and is used in the UI to show/hide the login button
 
-  $: if (accessToken) {
-    localStorage.setItem('gdrive_token', accessToken);
-  }
+  run(() => {
+    if (accessToken) {
+      localStorage.setItem('gdrive_token', accessToken);
+    }
+  });
 
   async function getFileSize(fileId: string): Promise<number> {
     try {
@@ -71,7 +77,11 @@
     }
   }
 
-  function xhrDownloadFileIdWithTracking(fileId: string, fileName: string, progressCallback: (loaded: number) => void) {
+  function xhrDownloadFileIdWithTracking(
+    fileId: string,
+    fileName: string,
+    progressCallback: (loaded: number) => void
+  ) {
     return new Promise<Blob>(async (resolve, reject) => {
       const { access_token } = gapi.auth.getToken();
       const xhr = new XMLHttpRequest();
@@ -113,7 +123,7 @@
       xhr.send();
     });
   }
-  
+
   // Keep this function for backward compatibility with other parts of the code
   function xhrDownloadFileId(fileId: string, fileName: string) {
     return xhrDownloadFileIdWithTracking(fileId, fileName, () => {});
@@ -127,7 +137,7 @@
     }
 
     accessToken = resp?.access_token;
-    
+
     const processId = 'connect-drive';
     progressTrackerStore.addProcess({
       id: processId,
@@ -141,7 +151,7 @@
         progress: 20,
         status: 'Checking for reader folder...'
       });
-      
+
       const { result: readerFolderRes } = await gapi.client.drive.files.list({
         q: `mimeType='application/vnd.google-apps.folder' and name='${READER_FOLDER}'`,
         fields: 'files(id)'
@@ -152,7 +162,7 @@
           progress: 40,
           status: 'Creating reader folder...'
         });
-        
+
         const { result: createReaderFolderRes } = await gapi.client.drive.files.create({
           resource: { mimeType: FOLDER_MIME_TYPE, name: READER_FOLDER },
           fields: 'id'
@@ -168,7 +178,7 @@
         progress: 60,
         status: 'Checking for volume data...'
       });
-      
+
       const { result: volumeDataRes } = await gapi.client.drive.files.list({
         q: `'${readerFolderId}' in parents and name='${VOLUME_DATA_FILE}'`,
         fields: 'files(id, name)'
@@ -182,7 +192,7 @@
         progress: 80,
         status: 'Checking for profiles...'
       });
-      
+
       const { result: profilesRes } = await gapi.client.drive.files.list({
         q: `'${readerFolderId}' in parents and name='${PROFILES_FILE}'`,
         fields: 'files(id, name)'
@@ -219,23 +229,23 @@
   export function logout() {
     // Remove token from localStorage
     localStorage.removeItem('gdrive_token');
-    
+
     // Clear the token from memory
     accessToken = '';
-    
+
     // Revoke the token with Google to ensure account picker shows up next time
     if (gapi.client.getToken()) {
       const token = gapi.client.getToken().access_token;
       // Clear the token from gapi client
       gapi.client.setToken(null);
-      
+
       // Revoke the token with Google's OAuth service
       fetch(`https://oauth2.googleapis.com/revoke?token=${token}`, {
         method: 'POST',
         headers: {
           'Content-Type': 'application/x-www-form-urlencoded'
         }
-      }).catch(error => {
+      }).catch((error) => {
         console.error('Error revoking token:', error);
       });
     }
@@ -280,7 +290,9 @@
   function createPicker() {
     // Create a view for ZIP/CBZ files
     const docsView = new google.picker.DocsView(google.picker.ViewId.DOCS)
-      .setMimeTypes('application/zip,application/x-zip-compressed,application/vnd.comicbook+zip,application/x-cbz')
+      .setMimeTypes(
+        'application/zip,application/x-zip-compressed,application/vnd.comicbook+zip,application/x-cbz'
+      )
       .setMode(google.picker.DocsViewMode.LIST)
       .setIncludeFolders(true)
       .setSelectFolderEnabled(true)
@@ -310,7 +322,7 @@
         fields: 'files(id, name, mimeType)',
         pageSize: 1000
       });
-      
+
       return result.files || [];
     } catch (error) {
       handleDriveError(error, 'listing files in folder');
@@ -322,16 +334,16 @@
   async function processFolder(folderId, folderName, scanProcessId) {
     const files = await listFilesInFolder(folderId);
     const allFiles = [];
-    
+
     // Process each file in the folder
     for (let i = 0; i < files.length; i++) {
       const file = files[i];
-      
+
       // Update the scan process with subfolder information
       progressTrackerStore.updateProcess(scanProcessId, {
-        status: `Scanning ${folderName} (${i+1}/${files.length}): ${file.name}`
+        status: `Scanning ${folderName} (${i + 1}/${files.length}): ${file.name}`
       });
-      
+
       if (file.mimeType === 'application/vnd.google-apps.folder') {
         // Recursively process subfolders - pass the same scanProcessId
         const subfolderFiles = await processFolder(file.id, file.name, scanProcessId);
@@ -341,25 +353,29 @@
         allFiles.push(file);
       }
     }
-    
+
     return allFiles;
   }
 
-  async function downloadAndProcessFiles(fileList: { id: string; name: string; mimeType: string }[], existingProcessId?: string) {
+  async function downloadAndProcessFiles(
+    fileList: { id: string; name: string; mimeType: string }[],
+    existingProcessId?: string
+  ) {
     // Import the worker pool dynamically
     const { WorkerPool } = await import('$lib/util/worker-pool');
-    
+
     // Use the existing processId if provided, otherwise create a new one
-    const overallProcessId = existingProcessId || 
+    const overallProcessId =
+      existingProcessId ||
       `download-batch-${Date.now()}-${Math.random().toString(36).substring(2, 9)}`;
-    
+
     // Sort files by name
     const sortedFiles = fileList.sort((a, b) => a.name.localeCompare(b.name));
-    
+
     // First, get the total size of all files to download
     let totalBytesToDownload = 0;
     let fileSizes: { [fileId: string]: number } = {};
-    
+
     // If we're using a new process ID (not continuing from scanning), create a new tracker
     if (!existingProcessId) {
       progressTrackerStore.addProcess({
@@ -378,12 +394,12 @@
         progress: 30
       });
     }
-    
+
     try {
       // Get file sizes in parallel using Promise.all
-      const sizePromises = sortedFiles.map(file => getFileSize(file.id));
+      const sizePromises = sortedFiles.map((file) => getFileSize(file.id));
       const sizes = await Promise.all(sizePromises);
-      
+
       // Store sizes in the map and calculate total
       sortedFiles.forEach((file, index) => {
         const size = sizes[index];
@@ -394,14 +410,14 @@
       console.error('Error calculating total size:', error);
       // Continue anyway with what we have
     }
-    
+
     // Update the progress tracker with the total size
     progressTrackerStore.updateProcess(overallProcessId, {
       status: `Preparing to download ${sortedFiles.length} files`,
       totalBytes: totalBytesToDownload,
       bytesLoaded: 0
     });
-    
+
     // Create a worker pool for parallel downloads
     // Use navigator.hardwareConcurrency to determine optimal number of workers
     // but limit to a reasonable number to avoid overwhelming the browser
@@ -410,36 +426,38 @@
     // This is not a hard limit - tasks that individually need more than 500MB can still run
     // It just prevents starting new tasks when the current pool already exceeds 500MB
     const memoryLimitMB = 500; // 500 MB memory threshold
-    console.log(`Creating worker pool with ${maxWorkers} workers and ${memoryLimitMB}MB memory threshold`);
+    console.log(
+      `Creating worker pool with ${maxWorkers} workers and ${memoryLimitMB}MB memory threshold`
+    );
     const workerPool = new WorkerPool(undefined, maxWorkers, memoryLimitMB);
-    
+
     // Track download progress
     const fileProgress: { [fileId: string]: number } = {};
     let completedFiles = 0;
     let failedFiles = 0;
     const processedFiles: { [fileId: string]: boolean } = {};
-    
+
     // Function to update overall progress
     const updateOverallProgress = () => {
       let totalLoaded = 0;
-      
+
       // Sum up progress from all files
       for (const fileId in fileProgress) {
         totalLoaded += fileProgress[fileId];
       }
-      
+
       // Calculate progress percentage
       let progressPercentage;
-      
+
       if (existingProcessId) {
         // If we're using an existing process ID, scale the download progress to 30-100%
         // Download phase is 30-100% of the total progress
-        progressPercentage = 30 + ((totalLoaded / totalBytesToDownload) * 70);
+        progressPercentage = 30 + (totalLoaded / totalBytesToDownload) * 70;
       } else {
         // If this is a standalone download, use the full 0-100% range
         progressPercentage = (totalLoaded / totalBytesToDownload) * 100;
       }
-      
+
       // Update the progress tracker
       progressTrackerStore.updateProcess(overallProcessId, {
         progress: progressPercentage,
@@ -447,43 +465,45 @@
         status: `Downloaded ${completedFiles} of ${sortedFiles.length} files (${failedFiles} failed)`
       });
     };
-    
+
     // Create a promise that resolves when all downloads are complete
     return new Promise<void>((resolve) => {
       // Function to check if all downloads are complete
       const checkAllComplete = () => {
         // Log current memory usage
         const memUsage = workerPool.memoryUsage;
-        console.log(`Memory usage: ${(memUsage.current / (1024 * 1024)).toFixed(2)}MB / ${(memUsage.max / (1024 * 1024)).toFixed(2)}MB (${memUsage.percentUsed.toFixed(2)}%)`);
-        
+        console.log(
+          `Memory usage: ${(memUsage.current / (1024 * 1024)).toFixed(2)}MB / ${(memUsage.max / (1024 * 1024)).toFixed(2)}MB (${memUsage.percentUsed.toFixed(2)}%)`
+        );
+
         if (completedFiles + failedFiles === sortedFiles.length) {
           // All files have been processed
           workerPool.terminate();
-          
+
           // Update the process to show completion
           progressTrackerStore.updateProcess(overallProcessId, {
             status: `All downloads complete (${failedFiles} failed)`,
             progress: 100,
             bytesLoaded: totalBytesToDownload
           });
-          
+
           // Only auto-remove the tracker if it's not part of a larger process
           if (!existingProcessId) {
             setTimeout(() => progressTrackerStore.removeProcess(overallProcessId), 3000);
           }
-          
+
           resolve();
         }
       };
-      
+
       // Add each file to the worker pool
       for (const fileInfo of sortedFiles) {
         // Initialize progress for this file
         fileProgress[fileInfo.id] = 0;
-        
+
         // Create a task for the worker pool
         // Estimate memory requirement based on file size
-        // We need memory for: 
+        // We need memory for:
         // 1. The downloaded file (fileSizes[fileInfo.id])
         // 2. Processing overhead (typically 2-3x the file size for decompression)
         const fileSize = fileSizes[fileInfo.id] || 0;
@@ -493,9 +513,11 @@
           fileSize * 3, // 3x file size for processing overhead
           50 * 1024 * 1024 // Minimum 50MB
         );
-        
-        console.log(`Adding task for ${fileInfo.name} with estimated memory requirement: ${(memoryRequirement / (1024 * 1024)).toFixed(2)}MB`);
-        
+
+        console.log(
+          `Adding task for ${fileInfo.name} with estimated memory requirement: ${(memoryRequirement / (1024 * 1024)).toFixed(2)}MB`
+        );
+
         workerPool.addTask({
           id: fileInfo.id,
           memoryRequirement,
@@ -516,67 +538,67 @@
                 dataSize: data.data.byteLength,
                 hasData: !!data.data
               });
-              
+
               // Create a Blob from the ArrayBuffer
               const blob = new Blob([data.data]);
               console.log(`Created blob of size ${blob.size} bytes`);
-              
+
               // Create a File object from the blob
               const file = new File([blob], data.fileName);
               console.log(`Created file object: ${file.name}, size: ${file.size} bytes`);
-              
+
               // Process the file
               await processFiles([file]);
               console.log(`Successfully processed file: ${file.name}`);
-              
+
               // Mark as completed
               completedFiles++;
               fileProgress[fileInfo.id] = fileSizes[fileInfo.id] || 0;
               processedFiles[fileInfo.id] = true;
-              
+
               updateOverallProgress();
               checkAllComplete();
             } catch (error) {
               console.error(`Error processing ${data.fileName}:`, error);
               showSnackbar(`Failed to process ${data.fileName}`);
-              
+
               // Mark as failed
               failedFiles++;
               processedFiles[fileInfo.id] = true;
-              
+
               updateOverallProgress();
               checkAllComplete();
             } finally {
-              releaseMemory()
+              releaseMemory();
             }
           },
           onError: (data) => {
             console.error(`Error downloading ${fileInfo.name}:`, data.error);
             showSnackbar(`Failed to download ${fileInfo.name}`);
-            
+
             // Mark as failed but count the size as downloaded for progress calculation
             failedFiles++;
             fileProgress[fileInfo.id] = fileSizes[fileInfo.id] || 0;
             processedFiles[fileInfo.id] = true;
-            
+
             updateOverallProgress();
             checkAllComplete();
           }
         });
       }
-      
+
       // If there are no files, resolve immediately
       if (sortedFiles.length === 0) {
         progressTrackerStore.updateProcess(overallProcessId, {
           status: 'No files to download',
           progress: 100
         });
-        
+
         // Only auto-remove the tracker if it's not part of a larger process
         if (!existingProcessId) {
           setTimeout(() => progressTrackerStore.removeProcess(overallProcessId), 3000);
         }
-        
+
         resolve();
       }
     });
@@ -586,12 +608,12 @@
     try {
       if (data[google.picker.Response.ACTION] == google.picker.Action.PICKED) {
         const docs = data[google.picker.Response.DOCUMENTS];
-        
+
         if (docs.length === 0) return;
-        
+
         // Create a unique ID for this entire process (scanning + downloading)
         const processId = `download-process-${Date.now()}-${Math.random().toString(36).substring(2, 9)}`;
-        
+
         // Create a single progress tracker for the entire process
         progressTrackerStore.addProcess({
           id: processId,
@@ -599,26 +621,26 @@
           progress: 0,
           status: 'Starting scan...'
         });
-        
+
         // Collect all files to download
         let allFiles = [];
-        
+
         try {
           // First, identify folders and regular files
           for (let i = 0; i < docs.length; i++) {
             const doc = docs[i];
-            
+
             // Update progress based on how many items we've processed
             // Scanning phase is 0-30% of the total progress
             const scanProgress = (i / docs.length) * 30;
-            
+
             if (doc.mimeType === 'application/vnd.google-apps.folder') {
               // Process folder to get all files inside
               progressTrackerStore.updateProcess(processId, {
                 progress: scanProgress,
                 status: `Scanning folder: ${doc.name}`
               });
-              
+
               // Pass the processId to processFolder
               const folderFiles = await processFolder(doc.id, doc.name, processId);
               allFiles.push(...folderFiles);
@@ -627,13 +649,13 @@
               allFiles.push(doc);
             }
           }
-          
+
           // Filter out any non-zip files that might have been included in folders
-          allFiles = allFiles.filter(file => {
+          allFiles = allFiles.filter((file) => {
             const mimeType = file.mimeType.toLowerCase();
             return mimeType.includes('zip') || mimeType.includes('cbz');
           });
-          
+
           if (allFiles.length === 0) {
             progressTrackerStore.updateProcess(processId, {
               progress: 100,
@@ -643,16 +665,16 @@
             showSnackbar('No compatible files found');
             return;
           }
-          
+
           // Update progress to show we're moving to download phase
           progressTrackerStore.updateProcess(processId, {
             progress: 30,
             status: `Found ${allFiles.length} files to download`
           });
-          
+
           // Download and process all files - pass the existing processId
           await downloadAndProcessFiles(allFiles, processId);
-          
+
           // After the entire process is complete, set a timeout to remove the progress tracker
           setTimeout(() => progressTrackerStore.removeProcess(processId), 3000);
         } catch (error) {
@@ -691,7 +713,7 @@
         progress: 50,
         status: 'Uploading...'
       });
-      
+
       const res = await uploadFile({
         accessToken,
         fileId: volumeDataId,
@@ -701,7 +723,7 @@
       });
 
       volumeDataId = res.id;
-      
+
       progressTrackerStore.updateProcess(processId, {
         progress: 100,
         status: 'Upload complete'
@@ -742,7 +764,7 @@
         progress: 50,
         status: 'Uploading...'
       });
-      
+
       const res = await uploadFile({
         accessToken,
         fileId: profilesId,
@@ -752,7 +774,7 @@
       });
 
       profilesId = res.id;
-      
+
       progressTrackerStore.updateProcess(processId, {
         progress: 100,
         status: 'Upload complete'
@@ -787,7 +809,7 @@
         progress: 50,
         status: 'Downloading...'
       });
-      
+
       const { body } = await gapi.client.drive.files.get({
         fileId: volumeDataId,
         alt: 'media'
@@ -807,7 +829,7 @@
         status: 'Download complete'
       });
       setTimeout(() => progressTrackerStore.removeProcess(processId), 3000);
-      
+
       showSnackbar('Volume data downloaded');
     } catch (error) {
       progressTrackerStore.updateProcess(processId, {
@@ -834,7 +856,7 @@
         progress: 50,
         status: 'Downloading...'
       });
-      
+
       const { body } = await gapi.client.drive.files.get({
         fileId: profilesId,
         alt: 'media'
@@ -854,7 +876,7 @@
         status: 'Download complete'
       });
       setTimeout(() => progressTrackerStore.removeProcess(processId), 3000);
-      
+
       showSnackbar('Profiles downloaded');
     } catch (error) {
       progressTrackerStore.updateProcess(processId, {
@@ -879,8 +901,9 @@
         <Button color="red" on:click={logout}>Log out</Button>
       </div>
       <p class="text-center">
-        Add your zipped manga files (ZIP or CBZ) to the <span class="text-primary-700">{READER_FOLDER}</span> folder
-        in your Google Drive.
+        Add your zipped manga files (ZIP or CBZ) to the <span class="text-primary-700"
+          >{READER_FOLDER}</span
+        > folder in your Google Drive.
       </p>
       <p class="text-center text-sm text-gray-500">
         You can select multiple ZIP/CBZ files or entire folders at once.
@@ -927,7 +950,7 @@
     <div class="flex justify-center pt-0 sm:pt-32">
       <button
         class="w-full border rounded-lg border-slate-600 p-10 border-opacity-50 hover:bg-slate-800 max-w-3xl"
-        on:click={signIn}
+        onclick={signIn}
       >
         <div class="flex sm:flex-row flex-col gap-2 items-center justify-center">
           <GoogleSolid size="lg" />
