@@ -10,12 +10,12 @@ module.exports = {
   plugins: ['@typescript-eslint'],
   parserOptions: {
     sourceType: 'module',
-    ecmaVersion: 2020,
+    ecmaVersion: 2022,
     extraFileExtensions: ['.svelte']
   },
   env: {
     browser: true,
-    es2017: true,
+    es2022: true,
     node: true
   },
   overrides: [
