@@ -1,24 +1,24 @@
-import { settings } from "$lib/settings";
-import { showSnackbar } from "$lib/util"
-import { get } from "svelte/store";
+import { settings } from '$lib/settings';
+import { showSnackbar } from '$lib/util';
+import { get } from 'svelte/store';
 
-export * from './cropper'
+export * from './cropper';
 
 export async function ankiConnect(action: string, params: Record<string, any>) {
   try {
     const res = await fetch('http://127.0.0.1:8765', {
       method: 'POST',
       body: JSON.stringify({ action, params, version: 6 })
-    })
-    const json = await res.json()
+    });
+    const json = await res.json();
 
     if (json.error) {
-      throw new Error(json.error)
+      throw new Error(json.error);
     }
 
     return json.result;
   } catch (e: any) {
-    showSnackbar(`Error: ${e?.message ?? e}`)
+    showSnackbar(`Error: ${e?.message ?? e}`);
   }
 }
 
@@ -30,11 +30,11 @@ export async function getCardInfo(id: string) {
 export async function getLastCardId() {
   const notesToday = await ankiConnect('findNotes', { query: 'added:1' });
   const id = notesToday.sort().at(-1);
-  return id
+  return id;
 }
 
 export async function getLastCardInfo() {
-  const id = await getLastCardId()
+  const id = await getLastCardId();
   return await getCardInfo(id);
 }
 
@@ -53,7 +53,7 @@ export async function blobToBase64(blob: Blob) {
 export async function imageToWebp(source: File) {
   const image = await createImageBitmap(source);
   const canvas = new OffscreenCanvas(image.width, image.height);
-  const context = canvas.getContext("2d");
+  const context = canvas.getContext('2d');
 
   if (context) {
     context.drawImage(image, 0, 0);
@@ -65,21 +65,16 @@ export async function imageToWebp(source: File) {
 }
 
 export async function updateLastCard(imageData: string | null | undefined, sentence?: string) {
-  const {
-    overwriteImage,
-    enabled,
-    grabSentence,
-    pictureField,
-    sentenceField
-  } = get(settings).ankiConnectSettings;
+  const { overwriteImage, enabled, grabSentence, pictureField, sentenceField } =
+    get(settings).ankiConnectSettings;
 
   if (!enabled) {
-    return
+    return;
   }
 
-  showSnackbar('Updating last card...', 10000)
+  showSnackbar('Updating last card...', 10000);
 
-  const id = await getLastCardId()
+  const id = await getLastCardId();
 
   if (getCardAgeInMin(id) >= 5) {
     showSnackbar('Error: Card created over 5 minutes ago');
@@ -93,7 +88,7 @@ export async function updateLastCard(imageData: string | null | undefined, sente
   }
 
   if (overwriteImage) {
-    fields[pictureField] = ''
+    fields[pictureField] = '';
   }
 
   if (imageData) {
@@ -104,15 +99,17 @@ export async function updateLastCard(imageData: string | null | undefined, sente
         picture: {
           filename: `_${id}.webp`,
           data: imageData.split(';base64,')[1],
-          fields: [pictureField],
-        },
-      },
-    }).then(() => {
-      showSnackbar('Card updated!')
-    }).catch((e) => {
-      showSnackbar(e)
+          fields: [pictureField]
+        }
+      }
     })
+      .then(() => {
+        showSnackbar('Card updated!');
+      })
+      .catch((e) => {
+        showSnackbar(e);
+      });
   } else {
-    showSnackbar('Something went wrong')
+    showSnackbar('Something went wrong');
   }
 }
