@@ -1,5 +1,5 @@
-import { describe, it, expect, vi } from 'vitest';
-import { render, fireEvent } from '@testing-library/svelte';
+import { describe, expect, it, vi } from 'vitest';
+import { fireEvent, render } from '@testing-library/svelte';
 import QuickAccess from '../QuickAccess.svelte';
 import { isReader } from '$lib/util';
 
@@ -15,16 +15,16 @@ describe('QuickAccess', () => {
     // Mock window.location
     const originalLocation = window.location;
     delete window.location;
-    window.location = { 
+    window.location = {
       href: '/series-uuid/volume-uuid',
       pathname: '/series-uuid/volume-uuid'
     } as any;
 
     const { getByText } = render(QuickAccess);
     const closeButton = getByText('Close reader');
-    
+
     await fireEvent.click(closeButton);
-    
+
     // Should navigate to /series-uuid
     expect(window.location.href).toBe('/series-uuid');
 
