@@ -1,3 +1,3 @@
-export * from './volume-data'
-export * from './settings'
-export * from './misc'
+export * from './volume-data';
+export * from './settings';
+export * from './misc';
