@@ -3,5 +3,5 @@ export function getItems(html: string) {
   const htmlDoc = parser.parseFromString(html, 'text/html');
 
   const items = htmlDoc.getElementsByTagName('a');
-  return [...items]
+  return [...items];
 }
