@@ -1,4 +1,4 @@
-import { describe, it, expect } from 'vitest';
+import { describe, expect, it } from 'vitest';
 import { countChars, getCharCount } from './count-chars';
 import type { Page } from '$lib/types';
 
