@@ -2,5 +2,5 @@ export * from './snackbar';
 export * from './upload';
 export * from './misc';
 export * from './modals';
-export * from './zip'
-export * from './cloud'
+export * from './zip';
+export * from './cloud';
