@@ -1,14 +1,14 @@
-import { page } from "$app/stores";
-import { get } from "svelte/store";
-import { showSnackbar } from "./snackbar";
-import { browser } from "$app/environment";
+import { page } from '$app/stores';
+import { get } from 'svelte/store';
+import { showSnackbar } from './snackbar';
+import { browser } from '$app/environment';
 
 export function clamp(num: number, min: number, max: number) {
   return Math.min(Math.max(num, min), max);
 }
 
 export function isReader() {
-  return get(page).route.id === '/[manga]/[volume]'
+  return get(page).route.id === '/[manga]/[volume]';
 }
 
 let timer: any;
@@ -42,12 +42,12 @@ type ExtaticPayload = {
   totalPages: number;
   currentLineCount: number;
   totalLineCount: number;
-}
+};
 
-type ExtaticEvent = 'mokuro-reader:page.change' | 'mokuro-reader:reader.closed'
+type ExtaticEvent = 'mokuro-reader:page.change' | 'mokuro-reader:reader.closed';
 
 export function fireExstaticEvent(event: ExtaticEvent, payload: ExtaticPayload) {
   if (browser) {
-    document.dispatchEvent(new CustomEvent(event, { detail: payload }))
+    document.dispatchEvent(new CustomEvent(event, { detail: payload }));
   }
 }
