@@ -34,25 +34,28 @@ interface ErrorMessage {
 async function downloadFile(fileId: string, fileName: string, accessToken: string) {
   try {
     // First get the file size
-    const sizeResponse = await fetch(`https://www.googleapis.com/drive/v3/files/${fileId}?fields=size`, {
-      headers: {
-        'Authorization': `Bearer ${accessToken}`
+    const sizeResponse = await fetch(
+      `https://www.googleapis.com/drive/v3/files/${fileId}?fields=size`,
+      {
+        headers: {
+          Authorization: `Bearer ${accessToken}`
+        }
       }
-    });
-    
+    );
+
     if (!sizeResponse.ok) {
       throw new Error(`Failed to get file size: ${sizeResponse.statusText}`);
     }
-    
+
     const sizeData = await sizeResponse.json();
     const totalSize = parseInt(sizeData.size, 10);
-    
+
     // Now download the file with progress tracking
     const xhr = new XMLHttpRequest();
     xhr.open('GET', `https://www.googleapis.com/drive/v3/files/${fileId}?alt=media`);
     xhr.setRequestHeader('Authorization', `Bearer ${accessToken}`);
     xhr.responseType = 'arraybuffer'; // Use arraybuffer instead of blob for better transferability
-    
+
     xhr.onprogress = (event) => {
       const progressMessage: ProgressMessage = {
         type: 'progress',
@@ -62,7 +65,7 @@ async function downloadFile(fileId: string, fileName: string, accessToken: strin
       };
       ctx.postMessage(progressMessage);
     };
-    
+
     xhr.onerror = () => {
       const errorMessage: ErrorMessage = {
         type: 'error',
@@ -71,7 +74,7 @@ async function downloadFile(fileId: string, fileName: string, accessToken: strin
       };
       ctx.postMessage(errorMessage);
     };
-    
+
     xhr.ontimeout = () => {
       const errorMessage: ErrorMessage = {
         type: 'error',
@@ -80,7 +83,7 @@ async function downloadFile(fileId: string, fileName: string, accessToken: strin
       };
       ctx.postMessage(errorMessage);
     };
-    
+
     xhr.onabort = () => {
       const errorMessage: ErrorMessage = {
         type: 'error',
@@ -89,7 +92,7 @@ async function downloadFile(fileId: string, fileName: string, accessToken: strin
       };
       ctx.postMessage(errorMessage);
     };
-    
+
     return new Promise<void>((resolve, reject) => {
       xhr.onload = () => {
         if (xhr.status >= 200 && xhr.status < 300) {
@@ -101,7 +104,7 @@ async function downloadFile(fileId: string, fileName: string, accessToken: strin
               responseSize: arrayBuffer.byteLength,
               status: xhr.status
             });
-            
+
             // Create a message with the ArrayBuffer
             const completeMessage: CompleteMessage = {
               type: 'complete',
@@ -109,12 +112,12 @@ async function downloadFile(fileId: string, fileName: string, accessToken: strin
               fileName,
               data: arrayBuffer
             };
-            
+
             console.log(`Worker: Sending complete message for ${fileName}`, {
               messageType: 'complete',
               dataSize: arrayBuffer.byteLength
             });
-            
+
             // Post the message with the ArrayBuffer as a transferable object
             ctx.postMessage(completeMessage, [arrayBuffer]);
             console.log(`Worker: Message posted for ${fileName}`);
@@ -139,7 +142,7 @@ async function downloadFile(fileId: string, fileName: string, accessToken: strin
           reject(new Error(`HTTP error ${xhr.status}: ${xhr.statusText}`));
         }
       };
-      
+
       xhr.send();
     });
   } catch (error) {
@@ -156,11 +159,11 @@ async function downloadFile(fileId: string, fileName: string, accessToken: strin
 // Listen for messages from the main thread
 ctx.addEventListener('message', async (event) => {
   console.log('Worker: Received message from main thread', event.data);
-  
+
   try {
     const { fileId, fileName, accessToken } = event.data as DownloadMessage;
     console.log(`Worker: Starting download for ${fileName} (${fileId})`);
-    
+
     await downloadFile(fileId, fileName, accessToken);
     console.log(`Worker: Download function completed for ${fileName}`);
   } catch (error) {
