@@ -6,7 +6,7 @@ const CACHE = `cache-${version}`;
 
 const ASSETS = [
   ...build, // the app itself
-  ...files  // everything in `static`
+  ...files // everything in `static`
 ];
 
 self.addEventListener('install', (event) => {
